changes around types, esp. User type to try to resolve errors in workflow
