--- src/clients/react-redux-vite/src/types/MSALTokenClaims.ts
@@ -1,9 +1,0 @@
-
-export interface MSALTokenClaims {
-  preferred_username?: string;
-  name?: string;
-  email?: string;
-  oid?: string;
-  tid?: string;
-  roles?: string[];
-}
