--- src/clients/react-redux-vite/src/types/User.ts
@@ -1,9 +1,0 @@
-
-export interface User {
-  id: string;
-  email: string;
-  name: string;
-  username?: string;
-  roles: string[];
-  tenantId?: string;
-}
